+++ changelog.md
@@ -17,6 +17,7 @@ Codion Change Log
 ### is.codion.swing.framework.ui
 - EntityApplicationPanel.Builder.applicationModelFactory() and applicationPanelFactory() renamed applicationModel() and applicationPanel().
 - EntityApplicationPanel.createLogControls() now protected final.
+- EntityTablePanel.conditions() renamed condition().
 
 ## 0.18.33
 ### is.codion.swing.common.ui